# apps/appointment-api-service/openapi.yaml
openapi: "3.1.0"
info:
  title: Keyloop — Multi-Tenant Vehicle Service Appointment API
  version: "1.0.0"
  description: |
    REST API for a multi-tenant vehicle service appointment system.
    All protected endpoints require a Bearer JWT in the Authorization header.
    All tenant-scoped endpoints require the `x-tenant-id` header matching the JWT `tenant_id` claim.
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XTenantId:
      name: x-tenant-id
      in: header
      required: true
      description: UUID of the target tenant. Must match the `tenant_id` claim in the JWT (unless isSuperAdmin).
      schema:
        type: string
        format: uuid
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    UserTenantDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        tenantName:
          type: string
        role:
          type: string
          enum: [Admin, TenantUser]
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
...
    TenantRequest:
      type: object
      required: [name, contactEmail]
      properties:
        name:
          type: string
          maxLength: 255
        contactEmail:
          type: string
          format: email
    TenantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        contactEmail:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Customer ──────────────────────────────────────────────────────────────
...
    VehicleRequest:
      type: object
      required: [customerId, licensePlate, make, model, year]
      properties:
        customerId:
          type: string
          format: uuid
        licensePlate:
          type: string
          maxLength: 20
        make:
          type: string
          maxLength: 100
        model:
          type: string
          maxLength: 100
        year:
          type: integer
          minimum: 1900
          maximum: 2100
    VehicleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        licensePlate:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        createdAt:
          type: string
          format: date-time

    # ── ServiceType ───────────────────────────────────────────────────────────
...
paths:
...
  /auth/tenants:
    get:
      tags: [Auth]
      summary: List tenants the current user belongs to
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserTenantDto'

  # ── Tenants (Admin only) ────────────────────────────────────────────────────
  /tenants:
    get:
      tags: [Tenants]
      summary: List all tenants (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TenantResponse'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        "403":
          description: Not an Admin

    post:
      tags: [Tenants]
      summary: Create a new tenant (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantRequest'
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        "403":
          description: Not an Admin
        "409":
          description: Tenant name already exists
  /tenants/{tenantId}:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Tenants]
      summary: Get a tenant by ID (Admin only)
      security:
        - BearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        "403": { description: Not an Admin }
        "404": { description: Tenant not found }
    put:
      tags: [Tenants]
      summary: Update a tenant (Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        "403": { description: Not an Admin }
        "404": { description: Tenant not found }
    delete:
      tags: [Tenants]
      summary: Soft-delete a tenant (Admin only)
      security:
        - BearerAuth: []
      responses:
        "204": { description: Tenant deactivated }
        "403": { description: Not an Admin }
        "404": { description: Tenant not found }
  /tenants/{tenantId}/users:
    parameters:
      - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Tenants]
      summary: Assign a role to a user in this tenant
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, role]
              properties:
                userId: { type: string, format: uuid }
                role: { type: string, enum: [TenantUser, TenantManager] }
      responses:
        "204": { description: Role assigned successfully }
        "403": { description: Forbidden }
  /tenants/{tenantId}/users/{userId}/role:
    parameters:
      - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Tenants]
      summary: Promote a user's role
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [TenantManager] }
      responses:
        "204": { description: Role promoted successfully }
        "403": { description: Forbidden }
  /tenants/{tenantId}/audit-logs:
    parameters:
      - { name: tenantId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Tenants]
      summary: Retrieve audit logs for a tenant
      security:
        - BearerAuth: []
      parameters:
        - { name: entityType, in: query, schema: { type: string } }
        - { name: entityId, in: query, schema: { type: string, format: uuid } }
        - { name: action, in: query, schema: { type: string } }
        - { name: userId, in: query, schema: { type: string, format: uuid } }
        - { name: startDate, in: query, schema: { type: string, format: date-time } }
        - { name: endDate, in: query, schema: { type: string, format: date-time } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: Audit logs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
        "403": { description: Forbidden }

  # ── Customers ───────────────────────────────────────────────────────────────
...
  /vehicles:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
    get:
      tags: [Vehicles]
      summary: List vehicles
      security:
        - BearerAuth: []
      parameters:
        - { name: customerId, in: query, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/VehicleResponse' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
    post:
      tags: [Vehicles]
      summary: Create a vehicle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleResponse'
        "422": { description: Customer does not belong to this tenant }
  /vehicles/{id}:
    parameters:
      - $ref: '#/components/parameters/XTenantId'
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Vehicles]
      summary: Get a vehicle
      security:
        - BearerAuth: []
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/VehicleResponse' } } } }
        "404": { description: Vehicle not found }
    put:
      tags: [Vehicles]
      summary: Update a vehicle
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequest'
      responses:
        "200": { content: { application/json: { schema: { $ref: '#/components/schemas/VehicleResponse' } } } }
    delete:
      tags: [Vehicles]
      summary: Soft-delete a vehicle
      security:
        - BearerAuth: []
      responses:
        "204": { description: Vehicle deleted }
        "409": { description: Vehicle referenced by active appointment }

  # ── ServiceTypes ─────────────────────────────────────────────────────────────
...